# docs/api-reference/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExportKit API",
    "version": "1.1.0",
    "description": "Production-ready data export API for SaaS teams. Generate CSV, JSON, and Excel exports with webhooks, signed URLs, email notifications, billing, team management, and scheduling.",
    "contact": {
      "name": "ExportKit Support",
      "email": "support@exportkit.io",
      "url": "https://exportkit.io"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.exportkit.io",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/api/jobs": {
      "post": {
        "summary": "Create export job",
        "description": "Create a new data export job (CSV, JSON, or Excel format). The job will be queued and processed asynchronously.",
        "tags": ["Jobs"],
        "security": [{ "ApiKeyAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateJobRequest" },
              "example": {
                "type": "csv",
                "payload": {
                  "table": "users",
                  "filters": { "status": "active" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Job created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateJobResponse" },
                "example": {
                  "id": "job_abc123",
                  "bullmqId": "bull:123",
                  "status": "QUEUED"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": {
                  "error": "Invalid request body",
                  "code": "VALIDATION_ERROR",
                  "details": [{ "path": ["type"], "message": "Required" }]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      },
      "get": {
        "summary": "List export jobs",
        "description": "List all export jobs for the authenticated customer with optional filtering and pagination.",
        "tags": ["Jobs"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": ["QUEUED", "PROCESSING", "COMPLETED", "FAILED"]
            },
            "description": "Filter by job status"
          },
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/PageSize" }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of jobs",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedJobList" }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/jobs/{id}": {
      "get": {
        "summary": "Get job status",
        "description": "Get detailed status of a specific export job by ID.",
        "tags": ["Jobs"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job status retrieved successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Job" },
                "examples": {
                  "queued": {
                    "summary": "Job is queued",
                    "value": {
                      "id": "job_abc123",
                      "status": "QUEUED",
                      "type": "csv",
                      "progress": 0,
                      "createdAt": "2024-01-15T09:00:00Z",
                      "updatedAt": "2024-01-15T09:00:00Z"
                    }
                  },
                  "completed": {
                    "summary": "Job completed",
                    "value": {
                      "id": "job_abc123",
                      "status": "COMPLETED",
                      "type": "csv",
                      "progress": 100,
                      "createdAt": "2024-01-15T09:00:00Z",
                      "updatedAt": "2024-01-15T09:05:00Z",
                      "result": {
                        "downloadUrl": "https://r2.exportkit.io/exports/job_abc123.csv",
                        "expiresAt": "2024-01-15T10:00:00Z",
                        "recordCount": 1500,
                        "fileSize": 45000,
                        "format": "csv",
                        "key": "exports/job_abc123.csv"
                      }
                    }
                  },
                  "failed": {
                    "summary": "Job failed",
                    "value": {
                      "id": "job_abc123",
                      "status": "FAILED",
                      "type": "csv",
                      "progress": 45,
                      "createdAt": "2024-01-15T09:00:00Z",
                      "updatedAt": "2024-01-15T09:02:00Z",
                      "error": {
                        "message": "Database connection timeout",
                        "code": "DB_TIMEOUT"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Job not found", "code": "JOB_NOT_FOUND" }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/jobs/{id}/download": {
      "get": {
        "summary": "Get download URL",
        "description": "Generate a fresh signed download URL for a completed export job. The URL expires after 1 hour. Returns 410 Gone if the file has expired and been deleted per the retention policy.",
        "tags": ["Jobs"],
        "security": [{ "ApiKeyAuth": [] }],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Download URL generated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DownloadUrlResponse" },
                "example": {
                  "downloadUrl": "https://r2.exportkit.io/exports/job_abc123.csv?X-Amz-Algorithm=...",
                  "expiresAt": "2024-01-15T10:30:00Z",
                  "fileExpiresAt": "2024-01-22T10:00:00Z"
                }
              }
            }
          },
          "400": {
            "description": "Export not yet complete",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Export not yet complete", "code": "EXPORT_NOT_READY" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "example": { "error": "Job not found", "code": "JOB_NOT_FOUND" }
              }
            }
          },
          "410": {
            "description": "Export file has expired and been deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string" },
                    "message": { "type": "string" },
                    "expiredAt": { "type": "string", "format": "date-time" }
                  }
                },
                "example": {
                  "error": "Gone",
                  "message": "Export file has expired and been deleted",
                  "expiredAt": "2024-01-22T09:00:00Z"
                }
              }
            }
          },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "Page": {
        "in": "query",
        "name": "page",
        "schema": { "type": "integer", "minimum": 1, "default": 1 },
        "description": "Page number (1-indexed)"
      },
      "PageSize": {
        "in": "query",
        "name": "pageSize",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 },
        "description": "Number of items per page (max 100)"
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Unauthorized — invalid or missing API key",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Unauthorized", "code": "UNAUTHORIZED" }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": { "type": "integer" },
            "description": "Seconds until the rate limit resets"
          },
          "X-RateLimit-Limit": {
            "schema": { "type": "integer" },
            "description": "Maximum requests allowed in the window"
          },
          "X-RateLimit-Remaining": {
            "schema": { "type": "integer" },
            "description": "Remaining requests in the current window"
          },
          "X-RateLimit-Reset": {
            "schema": { "type": "integer" },
            "description": "Unix timestamp when the rate limit window resets"
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Too many requests", "code": "RATE_LIMIT_EXCEEDED" }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Human-readable error message" },
          "code": { "type": "string", "description": "Machine-readable error code (UPPER_SNAKE_CASE)" },
          "details": {
            "type": "array",
            "items": { "type": "object" },
            "description": "Validation error details (when applicable)"
          }
        },
        "required": ["error", "code"]
      },
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "total": { "type": "integer", "description": "Total number of items" },
          "page": { "type": "integer", "description": "Current page number" },
          "pageSize": { "type": "integer", "description": "Items per page" },
          "hasNextPage": { "type": "boolean", "description": "Whether more pages exist" }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "description": "Unique job identifier" },
          "status": { "type": "string", "enum": ["QUEUED", "PROCESSING", "COMPLETED", "FAILED"] },
          "type": { "type": "string", "enum": ["csv", "json", "xlsx"] },
          "progress": { "type": "integer", "description": "Progress percentage (0-100)" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "result": {
            "type": "object",
            "properties": {
              "downloadUrl": { "type": "string" },
              "expiresAt": { "type": "string", "format": "date-time" },
              "recordCount": { "type": "integer" },
              "fileSize": { "type": "integer" },
              "format": { "type": "string" },
              "key": { "type": "string" }
            }
          },
          "error": {
            "type": "object",
            "properties": {
              "message": { "type": "string" },
              "code": { "type": "string" }
            }
          }
        }
      },
      "CreateJobRequest": {
        "type": "object",
        "required": ["type"],
        "properties": {
          "type": {
            "type": "string",
            "enum": ["csv", "json", "xlsx"],
            "description": "Export format type"
          },
          "payload": {
            "type": "object",
            "description": "Custom payload for the export job",
            "default": {}
          }
        }
      },
      "CreateJobResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "bullmqId": { "type": "string" },
          "status": { "type": "string", "enum": ["QUEUED", "PROCESSING", "COMPLETED", "FAILED"] }
        }
      },
      "PaginatedJobList": {
        "type": "object",
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } },
          "pagination": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      },
      "DownloadUrlResponse": {
        "type": "object",
        "properties": {
          "downloadUrl": { "type": "string", "description": "Signed URL for downloading the export file" },
          "expiresAt": { "type": "string", "format": "date-time", "description": "When the signed URL expires" },
          "fileExpiresAt": { "type": "string", "format": "date-time", "description": "When the file will be permanently deleted" }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Generate keys via the Dashboard or POST /api/keys endpoint."
      }
    }
  }
}